Created Tile.java and Hand.java
Created the two basic classes which will store all the information about tiles & the hands to hold them. Set up preliminary getters & setters, however Tile is a final class as it is immutable after creation. Also fixed a slight issue with the pom.xml throwing a warning that builds may be platform dependant.