ui: TileButton: add row/col properties to determine location on Table
The row and col properties allow us to determine its location on the
table.

In addition, there's a boolean property named originatesFromTable that
determines whether or not the TileButton exists on the table or exists
in the player's hand. It does this by binding to the row and column
properties. If the row and column values are greater than or equal to 0,
then we know it must exist on the table. If the TileButton stores a
negative integer for the row and column, then it must exist in the
player's hand.