backend: move playing AI turn in onEndTurnBtnClick
Previously, we were changing the player turn recursively in a
ChangeListener. This is considered bad practice by JavaFX and indeed it
caused mainly problems with the UI code. Thus, we run the AI turns on
clicking the end turn button.

Also, we introduce a delay in between each turn so that the user can see
what just happened. This should later be changed to using animations
instead, as imposing a static delay is considered bad user experience.