backend: use int instead of string for enum
Using string for the FXML option dropdown for player type selection
didn't work as we hoped, so we opt for another solution, where we do a
mapping from the index in the combo box to the enum's internal order.
This means that we must mirror the ordering of the ComboBox with the
enum Types in the OptionChoices.